in inverted index check old and new beacon format for backward compat
The beacon is used in the inverted index. This is because if you resolve
a ref you need to match against the beacon. Think of the following
pseudo-schema/object combination:

County(name=="Germany") -> hasCaptial -> City(name=="Berlin").

In Weaviate this would be roughly expressed as the following City:

properties: {
  "name": "Germany"
  "hasCaptial": [{
    "beacon": "weaviate://localhost/City/<id-of-Berlin>"
  }]
}

Now if a user where to filter for "Give me all the countries which a
capital named Berlin", this filter is a two step process.

1. Weaviate must identify all Cities where name=="Berlin". This is a
   simple inverted index lookup for the term "Berlin" on the City class.
   The result is a list of IDs (or a single ID in this case).

2. Now the user must find all Countries that link to the list of cities
   with the IDs from (1). This is yet another inverted index lookup. In
   this case the values for the "hasCaptial" inverted index bucket is
   the beacon format.

<-- This is the point where we get problems with two beacon formats.
Because we chose one at index time, and now we're using one at search
time. If they match, all is fine, but if they don't match, it's a
problem, as could be proven by the failing integration test that matched
for referenced values.

I was initially thinking, we could solve this by indexing both formats.
But this was a wrong line of thinking. We cannot do this because the old
format is only deprecated, so even on new imports we may end up with
both formats.

Instead I chose two look for both formats at search time. This is not
100% ideal because this introduces an additional cost. We should thus
aim to remove this backward-compatibility fix as soon as we can
guarantee that we will have no more old beacons, such as in v2.0.0.

Basically the fix now does a lookup in step two for two values chained
using an "OR" operation. Instead of looking for just one format, we are
now looking for

weaviate://localhost/City/<id-of-Berlin>
OR
weaviate://localhost/<id-of-Berlin>

Always one of the two should match.

Please also note the TODO-comment I added to the ref-filter integration
tests. This test currently only uses the old format, but I think it
would be nice to change it so it uses a mix of old and new formats, so
that we have a good guarantee that mixing formats works well.